mini/requirements.txt
openai>=1.12.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
tomli>=2.0.0
aiosqlite>=0.19.0
psutil>=5.9.0
